requirements/development.txt
asgiref==3.7.2
Django==4.2.4
django-axes==6.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
